+++ crates/bentley/src/daemon_logs.rs
@@ -1,0 +1,354 @@
+//! Daemon logging infrastructure for bentley
+//!
+//! This module provides persistent, structured logging for daemons with:
+//! - JSONL disk storage with unlimited capacity
+//! - Thread-safe async operations with internal locking
+//! - Optional console output (silent mode support) 
+//! - Full bentley macro integration for unified logging
+
+use serde::{Deserialize, Serialize};
+use chrono::{DateTime, Utc};
+
+// ============================================================================
+// TYPES AND DATA STRUCTURES
+// ============================================================================
+
+/// A structured log entry for daemon operations
+#[derive(Debug, Serialize, Deserialize, Clone)]
+pub struct LogEntry {
+  pub timestamp: DateTime<Utc>,
+  pub level: String,
+  pub message: String,
+  pub component: String,
+}
+
+/// Request structure for querying daemon logs
+#[derive(Debug, Serialize, Deserialize)]
+pub struct LogsRequest {
+  pub request: String,
+  pub limit: Option<usize>,
+  pub level: Option<String>,
+}
+
+/// Response structure for daemon log queries
+#[derive(Debug, Serialize, Deserialize)]
+pub struct LogsResponse {
+  pub success: bool,
+  pub logs: Vec<LogEntry>,
+  #[serde(skip_serializing_if = "Option::is_none")]
+  pub error: Option<ErrorInfo>,
+}
+
+/// Error information for daemon responses
+#[derive(Debug, Serialize, Deserialize)]
+pub struct ErrorInfo {
+  pub message: String,
+  pub tag: String,
+}
+
+/// Internal log storage implementation
+struct DaemonLogsInner {
+  log_file_path: std::path::PathBuf,
+  silent: bool,
+}
+
+/// Thread-safe disk-based log storage for daemons using JSONL format
+pub struct DaemonLogs {
+  inner: std::sync::Arc<tokio::sync::Mutex<DaemonLogsInner>>,
+}
+
+// ============================================================================
+// CONSTRUCTOR FUNCTIONS
+// ============================================================================
+
+impl LogsResponse {
+  pub fn success(logs: Vec<LogEntry>) -> Self {
+    Self { success: true, logs, error: None }
+  }
+
+  pub fn error(message: &str, tag: &str) -> Self {
+    Self {
+      success: false,
+      logs: Vec::new(),
+      error: Some(ErrorInfo {
+        message: message.to_string(),
+        tag: tag.to_string(),
+      }),
+    }
+  }
+}
+
+impl DaemonLogsInner {
+  /// Create a new daemon log storage that writes to the specified file path
+  fn new<P: AsRef<std::path::Path>>(log_file_path: P, silent: bool) -> std::io::Result<Self> {
+    let log_file_path = log_file_path.as_ref().to_path_buf();
+    
+    // Ensure parent directory exists
+    if let Some(parent) = log_file_path.parent() {
+      std::fs::create_dir_all(parent)?;
+    }
+    
+    // Create file if it doesn't exist (but don't truncate if it does)
+    if !log_file_path.exists() {
+      std::fs::File::create(&log_file_path)?;
+    }
+
+    Ok(Self { log_file_path, silent })
+  }
+
+  /// Add a log entry to storage (appends to JSONL file)
+  fn add_log(&mut self, level: &str, message: &str, component: &str) -> std::io::Result<()> {
+    let entry = LogEntry {
+      timestamp: Utc::now(),
+      level: level.to_string(),
+      message: message.to_string(),
+      component: component.to_string(),
+    };
+
+    // Serialize to JSON and append to file
+    let json_line = serde_json::to_string(&entry)
+      .map_err(|e| std::io::Error::new(std::io::ErrorKind::InvalidData, e))?;
+    
+    use std::fs::OpenOptions;
+    use std::io::Write;
+    
+    let mut file = OpenOptions::new()
+      .create(true)
+      .append(true)
+      .open(&self.log_file_path)?;
+    
+    writeln!(file, "{}", json_line)?;
+    file.flush()?;
+    
+    Ok(())
+  }
+
+  /// Retrieve logs with optional filtering and limiting (reads from JSONL file)
+  fn get_logs(&self, limit: Option<usize>, level_filter: Option<&str>) -> std::io::Result<Vec<LogEntry>> {
+    use std::fs::File;
+    use std::io::{BufRead, BufReader};
+
+    if !self.log_file_path.exists() {
+      return Ok(Vec::new());
+    }
+
+    let file = File::open(&self.log_file_path)?;
+    let reader = BufReader::new(file);
+    
+    let mut logs = Vec::new();
+    
+    for line_result in reader.lines() {
+      let line = line_result?;
+      if line.trim().is_empty() {
+        continue;
+      }
+      
+      match serde_json::from_str::<LogEntry>(&line) {
+        Ok(entry) => {
+          // Apply level filter
+          let matches_level = level_filter.map_or(true, |filter| {
+            filter == "all" || entry.level == filter
+          });
+          
+          if matches_level {
+            logs.push(entry);
+          }
+        },
+        Err(_) => {
+          // Skip malformed lines
+          continue;
+        }
+      }
+    }
+
+    // Sort by timestamp (newest first)
+    logs.sort_by(|a, b| b.timestamp.cmp(&a.timestamp));
+
+    // Apply limit
+    if let Some(limit) = limit {
+      logs.truncate(limit);
+    }
+
+    Ok(logs)
+  }
+
+  /// Get the path to the log file
+  fn log_file_path(&self) -> &std::path::Path {
+    &self.log_file_path
+  }
+
+  /// Check if the log file exists and has content
+  fn has_logs(&self) -> bool {
+    self.log_file_path.exists() && 
+      std::fs::metadata(&self.log_file_path)
+        .map(|m| m.len() > 0)
+        .unwrap_or(false)
+  }
+
+  /// Get the size of the log file in bytes
+  fn file_size(&self) -> std::io::Result<u64> {
+    let metadata = std::fs::metadata(&self.log_file_path)?;
+    Ok(metadata.len())
+  }
+}
+
+// ============================================================================
+// PUBLIC DAEMON LOGS API
+// ============================================================================
+
+impl DaemonLogs {
+  /// Create a new thread-safe daemon log storage
+  pub fn new<P: AsRef<std::path::Path>>(log_file_path: P) -> std::io::Result<Self> {
+    Self::new_with_silent(log_file_path, false)
+  }
+
+  /// Create a new thread-safe daemon log storage with silent option
+  pub fn new_with_silent<P: AsRef<std::path::Path>>(log_file_path: P, silent: bool) -> std::io::Result<Self> {
+    let inner = DaemonLogsInner::new(log_file_path, silent)?;
+    Ok(Self {
+      inner: std::sync::Arc::new(tokio::sync::Mutex::new(inner)),
+    })
+  }
+
+  /// Add a log entry (handles locking internally)
+  pub async fn add_log(&self, level: &str, message: &str, component: &str) -> std::io::Result<()> {
+    let mut guard = self.inner.lock().await;
+    guard.add_log(level, message, component)
+  }
+
+  /// Add a log entry (fire-and-forget, ignores errors)
+  pub async fn log(&self, level: &str, message: &str, component: &str) {
+    let _ = self.add_log(level, message, component).await;
+  }
+
+  /// Retrieve logs with optional filtering and limiting (handles locking internally)
+  pub async fn get_logs(&self, limit: Option<usize>, level_filter: Option<&str>) -> std::io::Result<Vec<LogEntry>> {
+    let guard = self.inner.lock().await;
+    guard.get_logs(limit, level_filter)
+  }
+
+  /// Get the path to the log file
+  pub async fn log_file_path(&self) -> std::path::PathBuf {
+    let guard = self.inner.lock().await;
+    guard.log_file_path().to_path_buf()
+  }
+
+  /// Check if the log file exists and has content
+  pub async fn has_logs(&self) -> bool {
+    let guard = self.inner.lock().await;
+    guard.has_logs()
+  }
+
+  /// Get the size of the log file in bytes
+  pub async fn file_size(&self) -> std::io::Result<u64> {
+    let guard = self.inner.lock().await;
+    guard.file_size()
+  }
+
+  /// Clone the DaemonLogs handle (cheap Arc clone)
+  pub fn clone(&self) -> Self {
+    Self {
+      inner: std::sync::Arc::clone(&self.inner),
+    }
+  }
+
+  /// Log an info message (to disk + console unless silent)
+  pub async fn info(&self, message: &str, component: &str) {
+    self.log("info", message, component).await;
+    
+    let guard = self.inner.lock().await;
+    if !guard.silent {
+      crate::info!(message);
+    }
+  }
+
+  /// Log a warning message (to disk + console unless silent)
+  pub async fn warn(&self, message: &str, component: &str) {
+    self.log("warn", message, component).await;
+    
+    let guard = self.inner.lock().await;
+    if !guard.silent {
+      crate::warn!(message);
+    }
+  }
+
+  /// Log a verbose message (to disk + console unless silent)
+  pub async fn verbose(&self, message: &str, component: &str) {
+    self.log("verbose", message, component).await;
+    
+    let guard = self.inner.lock().await;
+    if !guard.silent {
+      crate::verbose!(message);
+    }
+  }
+
+  /// Log an error message (to disk + console unless silent)
+  pub async fn error(&self, message: &str, component: &str) {
+    self.log("error", message, component).await;
+    
+    let guard = self.inner.lock().await;
+    if !guard.silent {
+      crate::error!(message);
+    }
+  }
+
+  /// Log a debug message (to disk + console unless silent)
+  pub async fn debug(&self, message: &str, component: &str) {
+    self.log("debug", message, component).await;
+    
+    let guard = self.inner.lock().await;
+    if !guard.silent {
+      crate::debug!(message);
+    }
+  }
+
+  /// Log a success message (to disk + console unless silent)
+  pub async fn success(&self, message: &str, component: &str) {
+    self.log("success", message, component).await;
+    
+    let guard = self.inner.lock().await;
+    if !guard.silent {
+      crate::success!(message);
+    }
+  }
+
+  /// Log an announcement message (to disk + console unless silent)
+  pub async fn announce(&self, message: &str, component: &str) {
+    self.log("announce", message, component).await;
+    
+    let guard = self.inner.lock().await;
+    if !guard.silent {
+      crate::announce!(message);
+    }
+  }
+
+  /// Log a spotlight message (to disk + console unless silent)
+  pub async fn spotlight(&self, message: &str, component: &str) {
+    self.log("spotlight", message, component).await;
+    
+    let guard = self.inner.lock().await;
+    if !guard.silent {
+      crate::spotlight!(message);
+    }
+  }
+
+  /// Log a flourish message (to disk + console unless silent)
+  pub async fn flourish(&self, message: &str, component: &str) {
+    self.log("flourish", message, component).await;
+    
+    let guard = self.inner.lock().await;
+    if !guard.silent {
+      crate::flourish!(message);
+    }
+  }
+
+  /// Log a showstopper message (to disk + console unless silent)
+  pub async fn showstopper(&self, message: &str, component: &str) {
+    self.log("showstopper", message, component).await;
+    
+    let guard = self.inner.lock().await;
+    if !guard.silent {
+      crate::showstopper!(message);
+    }
+  }
+}
